Fix compiler warnings for VS Win32 build (#79)
* 32 bit builds need explicit cast when converting uint64 to size_t

+++ Source/Model/COM/NMR_COMInterface_ModelWriter.cpp
@@ -213,7 +213,7 @@ namespace NMR {
 				throw CNMRException(NMR_ERROR_INSUFFICIENTBUFFERSIZE);
 
 			// TODO eliminate this copy, perhaps by allowing CExportStreamMemory to use existing buffers
-			memcpy(pBuffer, pStream->getData(), cbStreamSize);
+			memcpy(pBuffer, pStream->getData(), static_cast<size_t>(cbStreamSize));
 
 			return handleSuccess();
 		}
